webview: Turn js generation into a simple build system
Instead of just compiling the js file, now we do that but also
copy the files in `src/webview/assets/` to be imported from the
webview directly.

Now the compiled JS file is not a string that gets imported but
is a normal JS file.

The command to be run now is `yarn build-webview` - to better
reflect the process.