webview: Rework how dark theme styling is applied
Makes dark theme styling consistent with how this is done in the
web app. Instead of conditionally including `dark.css` do include
it always, but make the styles apply only when `body` has a class
of `night-mode`.

This makes the `code.css` file, already having styles in this
format work well in night mode.