Reactivate Reactotron!
Refresh our config and use of Reactotron in order to improve our
ability to debug the app.

This:
 * Updates dependencies to the latest Reactotron version
 * Simplifies the config, some of the `use` lines were not needed as
`useReactNative` already enabled them
 * enable again, but only for __DEV__
 * includes `__DEV__` as a ESLint global, preventing an error
 * turns the Reactotron `import` into a non-hoistable `require` and
moves it inside the `if` to not include in production builds
 * leaves `reactotronRedux` for a next commit, requires more changes

Note: To use Reactotron make sure the 'Debug JS Remotely' option is
selected inside your Emulator/Simulator/Device.