Use `progressiveTimeout` in `tryUntilSuccessful`
The retry strategy used in `tryUntilSuccessful` is suboptimal.
We can improve it and remove the logic completely if we reuse the
`progressiveTimeout` function.

The `maxRetries` and `timeoutMs` were not really used at this point
se do not try to reimplement this functionality.

Also a test that can test verifies the now unsupported and never
used `maxRetries = 0` is removed.