Logout if registering for events throws 4xx
Fixes #3253

If we encounter any 4xx error when trying to register for an event
queue, stop retrying and logout instead.

The only case currently known that will cause a 400 or 401 is when
the user has a no longer valid API key.

No 4xx error is fixable only by retrying the same request and failing
to register for an event queue is a critical error. Doing a logout
seems to be the right approach even in unforseen cases.